requirements.txt
streamlit
requests
psycopg2-binary
pandas
